+++ src/main/java/com/example/masterplanbbe/presentation/controller/ChatRestController.java
@@ -1,34 +1,22 @@
 package com.example.masterplanbbe.presentation.controller;
 
-import com.example.masterplanbbe.application.dto.ChatMessageDTO;
 import com.example.masterplanbbe.application.service.ChatService;
-import com.example.masterplanbbe.application.service.RedisPublisher;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
-import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/v1/chat")
 @RequiredArgsConstructor
-public class ChatController {
+public class ChatRestController {
 
-    private final RedisPublisher redisPublisher;
     private final ChatService chatService;
-
-    @MessageMapping("/chat")
-    public void sendMessage(ChatMessageDTO message) {
-        Long specId = message.getSpecId();
-        chatService.saveChatMessage(message);
-        redisPublisher.publish("spec:" + specId, message);
-    }
 
     @GetMapping("/recent")
     public ResponseEntity<?> getRecentMessages(@RequestParam Long specId) {
-        return ResponseEntity.ok()
-                .body(chatService.getRecentMessages(specId));
+        return ResponseEntity.ok().body(chatService.getRecentMessages(specId));
     }
 
     @GetMapping
@@ -38,7 +26,7 @@ public class ChatController {
         if (size <= 0) {
             return ResponseEntity.badRequest().body("사이즈는 0 보다 커야합니다.");
         }
-        Pageable pageable = PageRequest.of(0, Math.min(size, 100)); //최대 사이즈 100으로 제한
+        Pageable pageable = PageRequest.of(0, Math.min(size, 100)); // 최대 사이즈 100으로 제한
         return ResponseEntity.ok().body(chatService.getChatMessage(lastChatId, specId, pageable));
     }
